creating view and controller
