+++ ui_manager.py
@@ -1,0 +1,6 @@
+
+
+class UIManager:
+
+    def __init__(self):
+        pass
